Removed standalone use case and included it in Request
